Move SkipLayerNormalization shape inference into a norm submodule
Move the SkipLayerNormalization operator, plus its test, from ops.rs into
a new ops/norm.rs submodule, matching how this operator is grouped in
src/ops.rs.
